+++ package.json
@@ -20,7 +20,7 @@
     "@heroicons/react": "^2.2.0",
     "@next/third-parties": "^15.3.0",
     "cloudinary": "^2.6.0",
-    "@pharmatech/sdk": "^0.4.11",
+    "@pharmatech/sdk": "^0.4.13",
     "@radix-ui/react-slider": "^1.2.4",
     "@react-google-maps/api": "^2.20.6",
     "@types/google.maps": "^3.58.1",
