+++ src/app/(shop)/layout.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import { useState, ReactNode } from 'react';
+import { ReactNode } from 'react';
 import NavBar from '@/components/Navbar';
-import CartOverlay from '@/components/Cart/CartOverlay';
 import Footer from '@/components/Footer';
 
 type ShopLayoutProps = {
@@ -10,20 +9,16 @@ type ShopLayoutProps = {
 };
 
 export default function ShopLayout({ children }: ShopLayoutProps) {
-  const [isCartOpen, setIsCartOpen] = useState<boolean>(false);
-
   return (
     <div>
       <div className="relative bg-white">
         {/* Nav */}
         <div className="relative z-50">
-          <NavBar onCartClick={() => setIsCartOpen(true)} />
+          <NavBar />
         </div>
         {children}
       </div>
       <Footer />
-      {/* Cart Overlay */}
-      <CartOverlay isOpen={isCartOpen} closeCart={() => setIsCartOpen(false)} />
     </div>
   );
 }
