+++ src/components/Order/DeliveryInfo.tsx
@@ -1,34 +1,20 @@
 'use client';
 
-import React, { useEffect, useState } from 'react';
-import { useAuth } from '@/context/AuthContext';
-import { api } from '@/lib/sdkConfig';
+import React from 'react';
 import { Colors } from '@/styles/styles';
 import { OrderDetailedResponse } from '@pharmatech/sdk';
 
 interface Props {
-  orderId: string;
+  order: OrderDetailedResponse;
 }
 
-const DeliveryInfo: React.FC<Props> = ({ orderId }) => {
-  const { token } = useAuth();
-  const [orderDetail, setOrderDetail] = useState<OrderDetailedResponse | null>(
-    null,
-  );
-
-  useEffect(() => {
-    if (!orderId || !token) return;
-    api.order.getById(orderId, token).then(setOrderDetail).catch(console.error);
-  }, [orderId, token]);
-
-  const orderNumber = orderDetail?.id || 'No asignado';
-
+const DeliveryInfo: React.FC<Props> = ({ order }) => {
   const deliveryAddress =
-    orderDetail?.branch?.address ||
-    orderDetail?.details?.[0]?.productPresentation.product.name ||
+    order.branch?.address ||
+    order.details?.[0]?.productPresentation.product.name ||
     'No asignado';
 
-  const deliveryRecord = orderDetail?.orderDeliveries?.[0];
+  const deliveryRecord = order.orderDeliveries?.[0];
   const deliveryPersonName = deliveryRecord?.employee
     ? `${deliveryRecord.employee.firstName} ${deliveryRecord.employee.lastName}`
     : 'No asignado';
@@ -58,7 +44,7 @@ const DeliveryInfo: React.FC<Props> = ({ orderId }) => {
                 className="border px-4 py-2"
                 style={{ color: Colors.textMain }}
               >
-                {orderNumber}
+                {order.id.slice(0, 8)}
               </td>
             </tr>
             <tr>
